Exclude test 'TestPush_WebhookRequestURL' from race
